+++ easypy/units.py
@@ -146,9 +146,7 @@ class DataSize(int):
             return '0 bytes' if humanize else '0'
 
         if self in self.UNIT_NAMES:
-            unit = self.UNIT_NAMES[self]
-            return '1 {}'.format(unit) if humanize else unit
-
+            return self.UNIT_NAMES[self]
         for unit in self.SORTED_UNITS:
             name = self.UNIT_NAMES[unit]
             many = 'bytes' if name == 'byte' else name